Move eval module in a dedicated repo, break tests submodule
